refactor: move select functionality to a dedicated module for better organization
